requirements.txt
flask
flask-restful
openai
python-dotenv
llama-index
SQLAlchemy
mysql-connector-python